% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounteR.R
\docType{package}
\name{encounteR}
\alias{encounteR}
\alias{encounteR-package}
\title{Estimate animal density and abundance using random encounter models}
\description{
Random encounter models can be used to estimate animal abundances 
   and densities. They are particularly useful for studies using camera traps
   and acoustic detectors. They generally assume that animals move randomly
   in space. The expected nuber of encounters between the animal population 
   and the sensors can then be calculated. Given a certain numbe
}
\author{
Tim CD Lucas
}

